# pkg/models/localapi_swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Swagger CrowdSec
  description: CrowdSec local API
  contact:
    email: contact@crowdsec.net
host: 127.0.0.1
basePath: /v1
securityDefinitions:
  JWTAuthorizer:
    type: "apiKey"
    name: "Authorization: Bearer"
    in: "header"
  APIKeyAuthorizer:
    type: "apiKey"
    name: "X-Api-Key"
    in: "header"
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /decisions:
    get:
      description: Returns information about existing decisions
      summary: getDecisions
      tags:
        - bouncers
      operationId: getDecisions
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: scope
          in: query
          required: false
          type: string
          description: scope to which the decision applies (ie. IP/Range/Username/Session/...)
        - name: value
          in: query
          required: false
          type: string
          description: the value to match for in the specified scope
        - name: type
          in: query
          required: false
          type: string
          description: type of decision
        - name: ip
          in: query
          required: false
          type: string
          description: IP to search for (shorthand for scope=ip&value=)
        - name: range
          in: query
          required: false
          type: string
          description: range to search for (shorthand for scope=range&value=)
        - name: contains
          in: query
          required: false
          type: boolean
          description: indicate if you're looking for a decision that contains the value, or that is contained within the value
        - name: origins
          in: query
          required: false
          type: string
          description: 'Comma separated name of origins. If provided, then only the decisions originating from provided origins would be returned.'
        - name: scenarios_containing
          in: query
          required: false
          type: string
          description: 'Comma separated words. If provided, only the decisions created by scenarios containing any of the provided word would be returned.'
        - name: scenarios_not_containing
          in: query
          required: false
          type: string
          description: 'Comma separated words. If provided, only the decisions created by scenarios, not containing any of the provided word would be returned.'
      responses:
        '200':
          description: "successful operation"
          schema:
            $ref: '#/definitions/GetDecisionsResponse'
        '400':
          description: "400 response"
          schema:
            $ref: "#/definitions/ErrorResponse"
    head:
      description: Returns information about existing decisions
      summary: GetDecisions
      tags:
        - bouncers
      operationId: headDecisions
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: scope
          in: query
          required: false
          type: string
          description: scope to which the decision applies (ie. IP/Range/Username/Session/...)
        - name: value
          in: query
          required: false
          type: string
          description: the value to match for in the specified scope
        - name: type
          in: query
          required: false
          type: string
          description: type of decision
        - name: ip
          in: query
          required: false
          type: string
          description: IP to search for (shorthand for scope=ip&value=)
        - name: range
          in: query
          required: false
          type: string
          description: range to search for (shorthand for scope=range&value=)
        - name: contains
          in: query
          required: false
          type: boolean
          description: indicate if you're looking for a decision that contains the value, or that is contained within the value
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "400 response"
      security:
      - APIKeyAuthorizer: []
    delete:
      description: Delete decisions(s) for given filters (only from cscli)
      summary: deleteDecisions
      tags:
        - watchers
      operationId: deleteDecisions
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: scope
          in: query
          required: false
          type: string
          description: scope to which the decision applies (ie. IP/Range/Username/Session/...)
        - name: value
          in: query
          required: false
          type: string
          description: the value to match for in the specified scope
        - name: type
          in: query
          required: false
          type: string
          description: type of decision
        - name: ip
          in: query
          required: false
          type: string
          description: IP to search for (shorthand for scope=ip&value=)
        - name: range
          in: query
          required: false
          type: string
          description: range to search for (shorthand for scope=range&value=)
        - name: scenario
          in: query
          required: false
          type: string
          description: scenario to search
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/DeleteDecisionResponse'
          headers: {}
        '400':
          description: "400 response"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - JWTAuthorizer: []
...
definitions:
...
  Decision:
    title: Decision
    type: object
    properties:
      id:
        description: (only relevant for GET ops) the unique id
        type: integer
        readOnly: true
      uuid:
        description: 'only relevant for LAPI->CAPI, ignored for cscli->LAPI and crowdsec->LAPI'
        type: string
        readOnly: true
      origin:
        description: 'the origin of the decision : cscli, crowdsec'
        type: string
      type:
        description: 'the type of decision, might be ''ban'', ''captcha'' or something custom. Ignored when watcher (cscli/crowdsec) is pushing to APIL.'
        type: string
      scope:
        description: 'the scope of decision : does it apply to an IP, a range, a username, etc'
        type: string
      value:
        description: 'the value of the decision scope : an IP, a range, a username, etc'
        type: string
      duration:
        description: 'the duration of the decisions'
        type: string
      until:
        type: string
        description: 'the date until the decisions must be active'
      scenario:
        type: string
      simulated:
        type: boolean
        description: 'true if the decision result from a scenario in simulation mode'
        readOnly: true
    required:
      - origin
      - type
      - scope
      - value
      - duration
      - scenario
  DeleteDecisionResponse:
    title: DeleteDecisionResponse
    type: object
    properties:
      nbDeleted:
        type: string
        description: "number of deleted decisions"
...
  GetDecisionsResponse:
    title: GetDecisionsResponse
    type: array
    items:
      $ref: '#/definitions/Decision'
...
  ErrorResponse:
    type: "object"
    required:
    - "message"
    properties:
      message:
        type: "string"
        description: "Error message"
      errors:
        type: "string"
        description: "more detail on individual errors"
    title: "error response"
    description: "error response return by the API"
...